requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.3.5
litellm==1.10.1
boto3==1.28.0
azure-storage-blob==12.16.0
requests==2.31.0